### 2025-08-04 #8734 @editorialbot
[PRE REVIEW]: perbase: A performant per-base sequencing metrics toolkit with accurate handling of complex alignments
**Submitting author:** @sstadick (<a href="http://orcid.org/0009-0002-0915-9459">Seth Stadick</a>)
**Repository:** https://github.com/sstadick/perbase
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
