### 2025-08-13 57b8602
fix: resolve Ns when there is no ambiguity
