### 2025-08-13 bc4317e
fix: resolve Ns when there is no ambiguity (#87)
