### 2025-08-13 a3f853d
fix: update paper for 1.1 (#88)
