### 2025-08-13 c6809e9
chore: Release perbase version 1.1.0
