### 2025-09-22 7dce6f4
ci: remove `ubuntu-18.04` (#90)
