+++ .github/workflows/main.yml
@@ -15,9 +15,6 @@ jobs:
           - os: ubuntu-latest
             artifact_name: perbase
             asset_name: perbase-linux-amd64
-          - os: ubuntu-18.04
-            artifact_name: perbase
-            asset_name: perbase-linux-amd64-ubuntu_18_04
           - os: macos-latest
             artifact_name: perbase
             asset_name: perbase-macos-amd64
