### 2025-12-11 90e64d5
fix: allow empty seqeunce in BAM record in `perbase only-depth`
Handle empty sequences in BAM record when generating a pileup
to fix an out of bound error in `perbase only-depth` when a BAM record
does not have an associated sequence (`*` in sequence column in SAM
format).

Fixes: https://github.com/sstadick/perbase/issues/91
