### 2025-12-11 f5cfac8
test: add tests for empty SEQ handling, bump to v1.2.0
- Add 3 tests for empty SEQ records in base-depth pileup code path
- Bump version to 1.2.0
- Add changelog entry for PR #92
- Add README note about empty SEQ handling
