### 2025-12-11 af73bdc
chore: update deps and fixup lints
