+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0
 
 - Fix: Handle BAM records with empty SEQ fields (`*` in SAM format) in `base-depth`. Previously this would cause an out-of-bounds error. Now these reads still count toward depth, and their bases are counted as `N`. ([#92](https://github.com/sstadick/perbase/pull/92) by @ghuls)
+- Chore: Update deps and fixup lints
 
 ## 1.1.0
 
