+++ Cargo.toml
@@ -25,21 +25,20 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = "1.0.41"
-bio = "0.41"
+bio = "3.0.0"
 crossbeam = "0.8"
 crossbeam-channel = "0.5"
 csv = "1.1.6"
-env_logger = "0.9.0"
+env_logger = "0.11.8"
 grep-cli = "0.1.5"
-gzp = "0.10"
-itertools = "0.10.1"
+gzp = "2.0"
+itertools = "0.14.0"
 lazy_static = "1.4.0"
 log = "0.4.11"
 lru_time_cache = "0.11.1"
 num_cpus = "1.13.0"
 rayon = "1.4.0"
-#TODO add  features = ["libdeflate"] when https://github.com/rust-bio/rust-htslib/pull/341 is merged
-rust-htslib = { version = "0.39" }
+rust-htslib = {version = "0.51", default-features = false, features = ["libdeflate"]}
 rust-lapper = "1.0"
 serde = { version = "1.0.116", features = ["derive"] }
 smartstring = { version = "1.0.1", features = ["serde"] }
@@ -49,5 +48,5 @@ termcolor = "1.1.0"
 
 [dev-dependencies]
 proptest = "1.0.0"
-rstest = "0.11.0"
+rstest = "0.26.1"
 tempfile = "3.1.0"
