### 2025-12-11 4c14d18
chore: update deps and fixup lints
