### 2026-01-14 #8734 @fboehm
@apldx @sridhar0605 @ernfrid - would you like to review this submission for JOSS?
