### 2026-01-14 #9774 @editorialbot
[REVIEW]: perbase: A performant per-base sequencing metrics toolkit with accurate handling of complex alignments
**Submitting author:** @sstadick (<a href="http://orcid.org/0009-0002-0915-9459">Seth Stadick</a>)
**Repository:** https://github.com/sstadick/perbase
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.0
