### 2026-01-20 3c4fd4b
chore: update paper for 2026 JOSS reqs update
