### 2026-01-20 #9774 @sstadick
I have updated the paper to reflect the new JOSS reqs / address the missing sections above.
