### 2026-01-29 #9774 @chrisamiller
Overall, a really nice tool with improvements over existing packages. I expect that people will find it quite useful!

### Notes:
- installation was straightforward, tools runs well and seems to be performant.

- paper is well-written, succinct and informative

- there are a few typos in the README:
   "Calculate the only the depth at each base"
    - This section describes a three column file, but gives an example of a four column file:
```
"Or it can take files with three columns with headers that are like
<REF|chrom>\t<POS|chromStart>\t<END|chromEnd>\t<DEPTH|COV>"
```
(In general, the documentation is solid, but could use one more pass to discover typos like these.)

- The package is a little light on testing.  I certainly don't expect 100% coverage or anything, but a few more tests with tiny bams that run each sub-tool with at least the most common options might be a reasonable thing to add.

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

This seems to be missing.

### Running the tool

- If I give no inputs, the short help appears as expected, but it seems to be missing some of the options that appear in the full help (`-r, --ref-fasta` for example)

```./perbase-linux-amd64 base-depth
error: The following required arguments were not provided:
    <reads>

USAGE:
    perbase-linux-amd64 base-depth <reads> --channel-size-modifier <channel-size-modifier> --chunksize <chunksize> --compression-level <compression-level> --compression-threads <compression-threads> --exclude-flags <exclude-flags> --include-flags <include-flags> --mate-resolution-strategy <mate-resolution-strategy> --max-depth <max-depth> --min-mapq <min-mapq> --ref-cache-size <ref-cache-size> --threads <threads>
```
Not a dealbreaker, just unexpected.

- the `channel-size-modifier` was a bit confusing to me.  Why is it expressed in fraction of a gigabyte to allocate, instead of just specifying the mem usage in a more standard format (e.g. 1`00M` or `2G`)?

- I pretty quickly caused a few jobs to be killed by running out of RAM.  Would be nice to put a rough guideline for mem requirements in the docs, even if that's just something like "with default chunk sizes, estimated ## Gb per thread). If it's not that straightforward, that's fine too!

Great stuff - looking forward to integrating this into some of our workflows.
