### 2026-01-30 sstadick/perbase#96 @mbhall88
[JOSS review] cargo install dependency issue
This is part of https://github.com/openjournals/joss-reviews/issues/9774

---

When I try `cargo install` on my Mac M3 machine I get the following:

```
error: failed to run custom build command for `libz-ng-sys v1.1.23`

Caused by:
  process didn't exit successfully: `/var/folders/rn/cfgx24s15c7bjnzs777tplkw0000gn/T/cargo-install9rFmpu/release/build/libz-ng-sys-1ef5430c8f1536e9/build-script-cmake` (exit status: 101)
  --- stdout
  CMAKE_TOOLCHAIN_FILE_aarch64-apple-darwin = None
  CMAKE_TOOLCHAIN_FILE_aarch64_apple_darwin = None
  HOST_CMAKE_TOOLCHAIN_FILE = None
  CMAKE_TOOLCHAIN_FILE = None
  CMAKE_GENERATOR_aarch64-apple-darwin = None
  CMAKE_GENERATOR_aarch64_apple_darwin = None
  HOST_CMAKE_GENERATOR = None
  CMAKE_GENERATOR = None
  CMAKE_PREFIX_PATH_aarch64-apple-darwin = None
  CMAKE_PREFIX_PATH_aarch64_apple_darwin = None
  HOST_CMAKE_PREFIX_PATH = None
  CMAKE_PREFIX_PATH = None
  CMAKE_aarch64-apple-darwin = None
  CMAKE_aarch64_apple_darwin = None
  HOST_CMAKE = None
  CMAKE = None

  --- stderr
  running: cd "/var/folders/rn/cfgx24s15c7bjnzs777tplkw0000gn/T/cargo-install9rFmpu/release/build/libz-ng-sys-e717b44d496c7462/out/build" && CMAKE_PREFIX_PATH="" LC_ALL="C" "cmake" "/Users/michael/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/libz-ng-sys-1.1.23/src/zlib-ng" "-B" "/var/folders/rn/cfgx24s15c7bjnzs777tplkw0000gn/T/cargo-install9rFmpu/release/build/libz-ng-sys-e717b44d496c7462/out/build" "-DCMAKE_OSX_ARCHITECTURES=arm64" "-DBUILD_SHARED_LIBS=OFF" "-DZLIB_COMPAT=OFF" "-DZLIB_ENABLE_TESTS=OFF" "-DWITH_GZFILEOP=ON" "-DCMAKE_INSTALL_LIBDIR=lib" "-DCMAKE_INSTALL_PREFIX=/var/folders/rn/cfgx24s15c7bjnzs777tplkw0000gn/T/cargo-install9rFmpu/release/build/libz-ng-sys-e717b44d496c7462/out" "-DCMAKE_C_FLAGS= -ffunction-sections -fdata-sections -fPIC --target=arm64-apple-macosx -mmacosx-version-min=15.5 -w" "-DCMAKE_C_COMPILER=/usr/bin/cc" "-DCMAKE_CXX_FLAGS= -ffunction-sections -fdata-sections -fPIC --target=arm64-apple-macosx -mmacosx-version-min=15.5 -w" "-DCMAKE_CXX_COMPILER=/usr/bin/c++" "-DCMAKE_ASM_FLAGS= -ffunction-sections -fdata-sections -fPIC --target=arm64-apple-macosx -mmacosx-version-min=15.5 -w" "-DCMAKE_ASM_COMPILER=/usr/bin/cc" "-DCMAKE_BUILD_TYPE=Release"

  thread 'main' (2482481) panicked at /Users/michael/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/cmake-0.1.57/src/lib.rs:1132:5:

  failed to execute command: No such file or directory (os error 2)
  is `cmake` not installed?

  build script failed, must exit now
  note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
warning: build failed, waiting for other jobs to finish...
error: failed to compile `perbase v1.2.0`, intermediate artifacts can be found at `/var/folders/rn/cfgx24s15c7bjnzs777tplkw0000gn/T/cargo-install9rFmpu`.
To reuse those artifacts with a future compilation, set the environment variable `CARGO_TARGET_DIR` to that path.
```

It turns out I needed `cmake` installed on my system in order to install `perbase`. Please add this dependency to the installation docs.
