### 2026-01-30 sstadick/perbase#97 @mbhall88
[JOSS review] Panic when no index available
This is in relation to https://github.com/openjournals/joss-reviews/issues/9774

---
This is a small UX thing. When running `base-depth` with a BAM that is not indexed, there is a nice error message, but the program goes on to panic. Might be a little cleaner to not have a panic and catch this and exit cleanly?

```
$ perbase base-depth LRE_04.bam                                                                                                                                                                                                                     [main]
[2026-01-30T05:02:15Z INFO  perbase::commands::base_depth] Running base-depth on: "LRE_04.bam"
[2026-01-30T05:02:15Z INFO  perbase_lib::par_granges] Using 2 worker threads.
[2026-01-30T05:02:15Z INFO  perbase_lib::par_granges] Creating channel of length 3097332 (* 120 bytes to get mem)
[2026-01-30T05:02:15Z INFO  perbase_lib::par_granges] Reading from "LRE_04.bam"
[E::idx_find_and_load] Could not retrieve index file for 'LRE_04.bam'

thread '<unnamed>' (905494) panicked at /opt/conda/conda-bld/perbase_1765476947771/work/src/lib/par_granges.rs:160:72:
Indexed BAM/CRAM: BamInvalidIndex { target: "LRE_04.bam" }
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```
