### 2026-01-30 sstadick/perbase#98 @mbhall88
[JOSS review] Add contributing docs
This is in relation to https://github.com/openjournals/joss-reviews/issues/9774

---

It would be nice to have either a `CONTRIBUTING.md` file, or even just a small section in the README that outlines what things contributors should do prior to putting in any PRs — i.e. examples of how to running tests, lint, format, or any other quirks needed.
