### 2026-01-30 030a79e
fix: update system dependency list in readme
Closes https://github.com/sstadick/perbase/issues/96
