### 2026-01-30 d80dcde
fix: update system dependency list in readme
Closes https://github.com/sstadick/perbase/issues/96
