### 2026-01-30 c17cc7f
chore: add CONTRIBUTING guidelines
Addresses https://github.com/sstadick/perbase/issues/98
