### 2026-01-30 8ac8f01
chore: add CONTRIBUTING guidelines
Addresses https://github.com/sstadick/perbase/issues/98
