+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.0
 
 - [Fix](https://github.com/sstadick/perbase/pull/95): typo in `count_mate_resolutions` from @camlloyd 
+- [chore](https://github.com/sstadick/perbase/issues/98): added CONTRIBUTING.md per @mbhall88 suggestion
 
 ## 1.2.0
 
