### 2026-01-30 6eddee4
fix: handle missing index and ref files gracefully
fixes https://github.com/sstadick/perbase/issues/97
