### 2026-01-30 535d20b
fix: handle missing index and ref files gracefully
fixes https://github.com/sstadick/perbase/issues/97
