### 2026-02-02 sstadick/perbase#100 @chrisamiller
Typos in README
there are a few typos in the README:

```
Calculate the only the depth at each base"
```

This section describes a three column file, but gives an example of a four column file:
```
Or it can take files with three columns with headers that are like
<REF|chrom>\t<POS|chromStart>\t<END|chromEnd>\t<DEPTH|COV>
```

(In general, the documentation is solid, but could use one more pass to discover typos like these.)
