### 2026-02-02 sstadick/perbase#101 @chrisamiller
Short usage doesn't match full help
If I give no inputs, the short help appears as expected, but it seems to be missing some of the options that appear in the full help (-r, --ref-fasta for example)

```
$ perbase base-depth
error: The following required arguments were not provided:

USAGE:
    perbase-linux-amd64 base-depth <reads> --channel-size-modifier <channel-size-modifier> --chunksize <chunksize> --compression-level <compression-level> --compression-threads <compression-threads> --exclude-flags <exclude-flags> --include-flags <include-flags> --mate-resolution-strategy <mate-resolution-strategy> --max-depth <max-depth> --min-mapq <min-mapq> --ref-cache-size <ref-cache-size> --threads <threads>
```

If this common is a common convention, feel free to ignore, but it surprised me when looking for the fasta argument
