### 2026-02-02 #9774 @chrisamiller
Apologies for not creating issues for my review. I've done that now:
https://github.com/sstadick/perbase/issues/100
https://github.com/sstadick/perbase/issues/101
https://github.com/sstadick/perbase/issues/102
https://github.com/sstadick/perbase/issues/103
