### 2026-02-03 76e38e6
fix: updates from reviewer suggestions
