### 2026-02-04 bbdd202
fix: use `setting` instead of `global_settings` for AppSettings
The use of `global_settings` causes AppSettings to propagate to all
subcommands/substructs, which breaks positional argument parsing.

Signed-off-by: Rui Chen <rui@chenrui.dev>
