### 2026-02-04 60c99dd
chore: prep for v1.4.0 release
