+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "perbase"
-version = "1.3.0"
+version = "1.4.0"
 authors = ["Seth Stadick <sstadick@gmail.com>"]
 edition = "2024"
 license = "MIT"
