### 2026-02-04 sstadick/perbase#101 @sstadick
@chrisamiller - I've had to roll back this change due to https://github.com/sstadick/perbase/pull/104.

Just wanted to let you know in case it changes any evaluation! I don't think there's a way to get the sort of help message your describing at this time. I probably need to switch to clap over structopt and will leave this open as an enhancement.
