### 2026-02-05 #9774 @sstadick
- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **v1.4.0**
- [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. **https://doi.org/10.5281/zenodo.18496759**
