### 2026-02-09 #9774 @fboehm
@sstadick - is there a doi for hyperfine?
