### 2026-02-09 #9774 @sstadick
@fboehm there is not one that I have found, no. It's this tool https://github.com/sharkdp/hyperfine
