### 2026-02-14 #9774 @fboehm
@sstadick - The manuscript looks really good. I have no suggestions for improvement.

Comparing the archive and github repo: Author lists match; licenses (MIT) match; version tags (v1.4.0) match; titles don't match exactly.

It looks like the title of the archive is: Perbase
while the title of the manuscript is: perbase: A performant per-base sequencing metrics
toolkit with accurate handling of complex alignments

I suggest changing the archive title (and its capitalization) to match exactly the manuscript title.

thanks again!
