### 2026-02-19 #9774 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
