### 2026-03-15 #9774 @Kevin-Mattheus-Moerman
@sstadick As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Apologies for the delay in getting to this. Below are some final checks, some of which may require your attention:

👉 please add the city and country to your affiliation if possible. Please do not use acronyms for countries so spell them out in full.
- [x] Copyedit pass/checked for typos
