### 2026-03-16 #9774 @Kevin-Mattheus-Moerman
@sstadick perhaps this pull request fixes it: https://github.com/sstadick/perbase/pull/106
