### 2026-03-16 #9774 @Kevin-Mattheus-Moerman
@sstadick feel free to try to fix this issue on your own too if my attempt failed. You can call `@editorialbot generate pdf` to test changes made. More info on correct formatting here: https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain
