### 2026-03-16 e32915e
docs: link to paper
