+++ README.md
@@ -12,6 +12,8 @@ A highly parallelized utility for analyzing metrics at a per-base level.
 
 If a metric is missing, or performance is lacking. Please file a bug/feature ticket in issues.
 
+Check out the [paper](https://joss.theoj.org/papers/10.21105/joss.09774)!
+
 ## Why?
 
 Why `perbase` when so many other tools are out there? `perbase` leverages Rust's concurrency system to automagically parallelize over your input regions. This leads to orders of magnitude faster runtimes that scale with the compute resources that you have available. Additionally, `perbase` aims to be more accurate than other tools. E.g.: `perbase` counts DELs toward depth, `bam-readcount` does not, `perbase` does not count REF_SKIPs toward depth, `sambamba` does.
@@ -35,7 +37,6 @@ The build scripts of some dependencies will need to compile `c` libraries.
 - `cmake`
 - `gcc`
 
-
 ## Tools
 
 ### base-depth
@@ -46,29 +47,29 @@ The `base-depth` tool walks over every position in the BAM/CRAM file and calcula
 
 The output columns are as follows:
 
-| Column         | Description                                                                                        |
-| -------------- | -------------------------------------------------------------------------------------------------- |
-| REF            | The reference sequence name                                                                        |
-| POS            | The position on the reference sequence                                                             |
-| REF_BASE       | The reference base at the position, column excluded if no reference was supplied                   |
-| DEPTH          | The total depth at the position SUM(A, C, T, G, N, R, Y, S, W, K, M, DEL)                          |
-| A              | Total A nucleotides seen at this position                                                          |
-| C              | Total C nucleotides seen at this position                                                          |
-| G              | Total G nucleotides seen at this position                                                          |
-| T              | Total T nucleotides seen at this position                                                          |
-| N              | Total N nucleotides seen at this position                                                          |
-| R              | Total R nucleotides seen at this position                                                          |
-| Y              | Total Y nucleotides seen at this position                                                          |
-| S              | Total S nucleotides seen at this position                                                          |
-| W              | Total W nucleotides seen at this position                                                          |
-| K              | Total K nucleotides seen at this position                                                          |
-| M              | Total M nucleotides seen at this position                                                          |
-| INS            | Total insertions that start at the base to the right of this position                              |
-| DEL            | Total deletions covering this position                                                             |
-| REF_SKIP       | Total reference skip operations covering this position                                             |
-| FAIL           | Total reads failing filters that covered this position (their bases were not counted toward depth) |
-| COUNT_OF_MATE_RESOLUTIONS | Total number times that mate resolution needed to be done                               |
-| NEAR_MAX_DEPTH | Flag to indicate if this position came within 1% of the max depth specified                        |
+| Column                    | Description                                                                                        |
+| ------------------------- | -------------------------------------------------------------------------------------------------- |
+| REF                       | The reference sequence name                                                                        |
+| POS                       | The position on the reference sequence                                                             |
[…]
